Move all data structures to separate include folder
