+++ inc/dataStructures/intervalTree-private.hpp
