+++ tests/QueueTest.cpp
@@ -1,4 +1,4 @@
-#include "sim/threadsafeQueue.hpp"
+#include "dataStructures/threadsafeQueue.hpp"
 #include <gtest/gtest.h>
 #include <chrono>
 #include <thread>
